requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
alembic==1.6.5
anyio==3.3.0
asgiref==3.4.1
async-timeout==3.0.1
bcrypt==3.2.0
blinker==1.4
boto3==1.18.22
botocore==1.21.22
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
cssselect==1.1.0
cssutils==2.3.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
emails==0.6
fakeredis==1.6.0
fastapi==0.67.0
fastapi-mail==0.4.1
greenlet==1.1.0
gunicorn==20.0.4
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==3.2
Jinja2==3.0.1
jmespath==0.10.0
jwt==1.2.0
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
motor==2.5.0
mysqlclient==2.0.3
newrelic==6.8.1.164
passlib==1.7.4
premailer==3.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
redis==3.5.3
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.22
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
